Rename [uuid] wildcard
