+++ src/bot/buttons/ManageIcon.ts
@@ -3,7 +3,7 @@ import Button from "../structs/Button";
 import players from "../../database/schemas/players";
 import { colors } from "../bot";
 import { constantCase } from "change-case";
-import { getCustomIconUrl } from "../../routes/players/:uuid/icon";
+import { getCustomIconUrl } from "../../routes/players/[uuid]/icon";
 import { Permission } from "../../types/Permission";
 import { GlobalIcon } from "../../types/GlobalIcon";
