+++ src/database/schemas/players.ts
@@ -3,7 +3,7 @@ import { client, fetchGuild } from "../../bot/bot";
 import Logger from "../../libs/Logger";
 import { GuildMember } from "discord.js";
 import { constantCase } from "change-case";
-import { generateSecureCode } from "../../routes/players/:uuid/connections";
+import { generateSecureCode } from "../../routes/players/[uuid]/connections";
 import { config } from "../../libs/config";
 import { Permission } from "../../types/Permission";
 import { getCachedRoles, Role } from "./roles";
