+++ src/index.ts
@@ -22,7 +22,7 @@ import cors from "@elysiajs/cors";
 import { verify as verifyMailOptions } from "./libs/mailer";
 import { getLatestCommit, retrieveData } from "./libs/git-commit-data";
 import { startEntitlementExpiry, startMetrics, startReferralReset, startRoleCacheJob } from "./libs/cron-jobs";
-import { formatUUID } from "./routes/players/:uuid/root";
+import { formatUUID } from "./routes/players/[uuid]/root";
 import { config } from "./libs/config";
 import { join } from "path";
