+++ src/libs/discord-notifier.ts
@@ -1,7 +1,7 @@
 import * as bot from "../bot/bot";
 import { ActionRowBuilder, ButtonBuilder, ButtonStyle, EmbedBuilder, TextChannel } from "discord.js";
 import { Profile } from "./mojang";
-import { getCustomIconUrl } from "../routes/players/:uuid/icon";
+import { getCustomIconUrl } from "../routes/players/[uuid]/icon";
 import { capitalCase } from "change-case";
 import { config } from "./config";
