+++ src/routes/players/[uuid]/ban.ts
