+++ src/routes/players/[uuid]/icon.ts
