+++ src/routes/players/[uuid]/position.ts
