+++ src/routes/players/[uuid]/referral.ts
