+++ src/routes/players/[uuid]/root.ts
